[config] move config.Chain to blockchain package
