+++ blockchain/filedao/testing.go
@@ -22,12 +22,13 @@ import (
 	"github.com/iotexproject/iotex-core/action"
 	"github.com/iotexproject/iotex-core/blockchain/block"
 	"github.com/iotexproject/iotex-core/blockchain/genesis"
-	"github.com/iotexproject/iotex-core/config"
 	"github.com/iotexproject/iotex-core/db"
 	"github.com/iotexproject/iotex-core/db/batch"
 	"github.com/iotexproject/iotex-core/test/identityset"
 	"github.com/iotexproject/iotex-core/testutil"
 )
+
+var _defaultEVMNetworkID uint32 = 4689
 
 type (
 	// testInMemFd is an in-memory FileDAO
@@ -69,7 +70,7 @@ func newFileDAOv2InMem(bottom uint64) (*fileDAOv2, error) {
 		blkCache: cache.NewThreadSafeLruCache(16),
 		kvStore:  db.NewMemKVStore(),
 		batch:    batch.NewBatch(),
-		deser:    block.NewDeserializer(config.Default.Chain.EVMNetworkID),
+		deser:    block.NewDeserializer(_defaultEVMNetworkID),
 	}
 	return &fd, nil
 }
