+++ blockchain/integrity/testcontract_test.go
@@ -9,6 +9,6 @@ package integrity
 import "encoding/hex"
 
 var (
-	_constantinopleOpCodeContract, _ = hex.DecodeString("608060405234801561001057600080fd5b506104d5806100206000396000f3fe608060405234801561001057600080fd5b50600436106100885760003560e01c806381ea44081161005b57806381ea440814610101578063a91b336214610159578063c2bc2efc14610177578063f5eacece146101cf57610088565b80635bec9e671461008d57806360fe47b1146100975780636bc8ecaa146100c5578063744f5f83146100e3575b600080fd5b6100956101ed565b005b6100c3600480360360208110156100ad57600080fd5b8101908080359060200190929190505050610239565b005b6100cd610270565b6040518082815260200191505060405180910390f35b6100eb6102b3565b6040518082815260200191505060405180910390f35b6101436004803603602081101561011757600080fd5b81019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506102f6565b6040518082815260200191505060405180910390f35b61016161036a565b6040518082815260200191505060405180910390f35b6101b96004803603602081101561018d57600080fd5b81019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506103ad565b6040518082815260200191505060405180910390f35b6101d761045f565b6040518082815260200191505060405180910390f35b5b60011561020b5760008081548092919060010191905055506101ee565b7f8bfaa460932ccf8751604dd60efa3eafa220ec358fccb32ef703f91c509bc3ea60405160405180910390a1565b80600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a250565b6000805460081d905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b6000805460081c905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b60008073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141561033157600080fd5b813f9050807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a2809050919050565b6000805460081b905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b60008073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614156103e857600080fd5b7fbde7a70c2261170a87678200113c8e12f82f63d0a1d1cfa45681cbac328e87e382600054604051808373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020018281526020019250505060405180910390a16000549050919050565b60008080602060406000f59150817fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a2819150509056fea265627a7a72305820209a8ef04c4d621759f34878b27b238650e8605c8a71d6efc619a769a64aa9cc64736f6c634300050a0032")
 	_codeStoreOutOfGasContract, _    = hex.DecodeString("60806040526040516200332738038062003327833981810160405260808110156200002957600080fd5b8101908080519060200190929190805190602001909291908051906020019092919080519060200190929190505050838383836200006c620004c860201b60201c565b8411620000c5576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526026815260200180620033016026913960400191505060405180910390fd5b6000620000d885620004d960201b60201c565b9050620000eb81620004f860201b60201c565b62000142576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252602d8152602001806200328b602d913960400191505060405180910390fd5b803410156200019d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526022815260200180620032b86022913960400191505060405180910390fd5b42620001ae620007e960201b60201c565b84031162000208576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401808060200182810382526027815260200180620032da6027913960400191505060405180910390fd5b806003819055508260058190555081600260006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555083600681905550336000806101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055506000349050600354810390506000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167f327ca9276da9073b583440165cb887319e7aeaf4003f14e92c1bbeee913e9b9c6003546040518082815260200191505060405180910390a26000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801562000383573d6000803e3d6000fd5b506000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167fb54ecacfeaef46d29d83a7ee9ffaae207283e8a0f5df648dba2d52a1454e489e826040518082815260200191505060405180910390a26000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167f53ddc5f55aac282e58e964d06a34c46ab814f1ea04d4ccf6eed6df8ba36d3f5e600354600654604051808381526020018281526020019250505060405180910390a26000600760016101000a81548160ff0219169083600e8111156200048e57fe5b02179055506000600760006101000a81548160ff02191690836004811115620004b357fe5b02179055505050505050505050505062000a56565b6000683635c9adc5dea00000905090565b6000680ad78ebc5ac6200000808381620004ef57fe5b04029050919050565b6000806000905062000512836000620007f460201b60201c565b81101562000530576200052d836000620007f460201b60201c565b90505b62000543836001620007f460201b60201c565b81101562000561576200055e836001620007f460201b60201c565b90505b62000574836002620007f460201b60201c565b81101562000592576200058f836002620007f460201b60201c565b90505b620005a5836003620007f460201b60201c565b811015620005c357620005c0836003620007f460201b60201c565b90505b620005d6836004620007f460201b60201c565b811015620005f457620005f1836004620007f460201b60201c565b90505b62000607836005620007f460201b60201c565b811015620006255762000622836005620007f460201b60201c565b90505b62000638836006620007f460201b60201c565b811015620006565762000653836006620007f460201b60201c565b90505b62000669836007620007f460201b60201c565b811015620006875762000684836007620007f460201b60201c565b90505b6200069a836008620007f460201b60201c565b811015620006b857620006b5836008620007f460201b60201c565b90505b620006cb836009620007f460201b60201c565b811015620006e957620006e6836009620007f460201b60201c565b90505b620006fc83600a620007f460201b60201c565b8110156200071a576200071783600a620007f460201b60201c565b90505b6200072d83600b620007f460201b60201c565b8110156200074b576200074883600b620007f460201b60201c565b90505b6200075e83600c620007f460201b60201c565b8110156200077c576200077983600c620007f460201b60201c565b90505b6200078f83600d620007f460201b60201c565b811015620007ad57620007aa83600d620007f460201b60201c565b90505b620007c083600e620007f460201b60201c565b811015620007de57620007db83600e620007f460201b60201c565b90505b828114915050919050565b600062015180905090565b60006008600e8111156200080457fe5b82600e8111156200081157fe5b14806200083657506009600e8111156200082757fe5b82600e8111156200083457fe5b145b806200085a5750600a600e8111156200084b57fe5b82600e8111156200085857fe5b145b806200087e5750600b600e8111156200086f57fe5b82600e8111156200087c57fe5b145b80620008a25750600c600e8111156200089357fe5b82600e811115620008a057fe5b145b80620008c65750600d600e811115620008b757fe5b82600e811115620008c457fe5b145b80620008e95750600e80811115620008da57fe5b82600e811115620008e757fe5b145b15620009035760018381620008fa57fe5b04905062000a50565b6004600e8111156200091157fe5b82600e8111156200091e57fe5b14156200093957600883816200093057fe5b04905062000a50565b6005600e8111156200094757fe5b82600e8111156200095457fe5b14156200096f57600683816200096657fe5b04905062000a50565b6006600e8111156200097d57fe5b82600e8111156200098a57fe5b1415620009a557600383816200099c57fe5b04905062000a50565b6007600e811115620009b357fe5b82600e811115620009c057fe5b1415620009db5760028381620009d257fe5b04905062000a50565b6002600e811115620009e957fe5b82600e811115620009f657fe5b141562000a11576001838162000a0857fe5b04905062000a50565b6000600e81111562000a1f57fe5b82600e81111562000a2c57fe5b141562000a4b5762000a43620004c860201b60201c565b905062000a50565b600090505b92915050565b6128258062000a666000396000f3fe6080604052600436106101355760003560e01c8063996b6fc1116100ab578063b7213b521161006f578063b7213b5214610460578063c6ee20d2146104af578063c9744029146104e8578063cea1295914610513578063d7606c4114610541578063fa6642c61461059d57610135565b8063996b6fc11461034a578063a307afda14610375578063a81af5f7146103a0578063a89ae4ba146103cb578063ab7695121461042257610135565b80632b68bb2d116100fd5780632b68bb2d1461021c5780633e9ee9e5146102265780635e4ff9b7146102515780636f09caed1461028f5780637150d8ae146102c857806392e4a9ea1461031f57610135565b806308551a531461013a5780630a9020271461019157806311da60b4146101bc578063142dfaa6146101c65780631856c845146101f1575b600080fd5b34801561014657600080fd5b5061014f6105f0565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561019d57600080fd5b506101a6610615565b6040518082815260200191505060405180910390f35b6101c461061b565b005b3480156101d257600080fd5b506101db610b52565b6040518082815260200191505060405180910390f35b3480156101fd57600080fd5b50610206610b5d565b6040518082815260200191505060405180910390f35b610224610b63565b005b34801561023257600080fd5b5061023b610d48565b6040518082815260200191505060405180910390f35b34801561025d57600080fd5b5061028d6004803603602081101561027457600080fd5b81019080803560ff169060200190929190505050610d53565b005b34801561029b57600080fd5b506102a4610f88565b6040518082600e8111156102b457fe5b60ff16815260200191505060405180910390f35b3480156102d457600080fd5b506102dd610f9b565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561032b57600080fd5b50610334610fc1565b6040518082815260200191505060405180910390f35b34801561035657600080fd5b5061035f610fcc565b6040518082815260200191505060405180910390f35b34801561038157600080fd5b5061038a610fd7565b6040518082815260200191505060405180910390f35b3480156103ac57600080fd5b506103b5610fdd565b6040518082815260200191505060405180910390f35b3480156103d757600080fd5b506103e0610fe3565b604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390f35b34801561042e57600080fd5b5061045e6004803603602081101561044557600080fd5b81019080803560ff169060200190929190505050611009565b005b34801561046c57600080fd5b506104996004803603602081101561048357600080fd5b810190808035906020019092919050505061121a565b6040518082815260200191505060405180910390f35b3480156104bb57600080fd5b506104c4611238565b604051808260048111156104d457fe5b60ff16815260200191505060405180910390f35b3480156104f457600080fd5b506104fd61124b565b6040518082815260200191505060405180910390f35b61053f6004803603602081101561052957600080fd5b810190808035906020019092919050505061125c565b005b34801561054d57600080fd5b506105876004803603604081101561056457600080fd5b8101908080359060200190929190803560ff1690602001909291905050506116c6565b6040518082815260200191505060405180910390f35b3480156105a957600080fd5b506105d6600480360360208110156105c057600080fd5b81019080803590602001909291905050506118ec565b604051808215151515815260200191505060405180910390f35b6000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60035481565b42610624610fcc565b600554011061067e576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260438152602001806127ae6043913960600191505060405180910390fd5b6002600481111561068b57fe5b600760009054906101000a900460ff1660048111156106a657fe5b14806106d75750600160048111156106ba57fe5b600760009054906101000a900460ff1660048111156106d557fe5b145b61072c576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252602e815260200180612716602e913960400191505060405180910390fd5b6000600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905060008173ffffffffffffffffffffffffffffffffffffffff16638da5cb5b6040518163ffffffff1660e01b815260040160206040518083038186803b15801561079b57600080fd5b505afa1580156107af573d6000803e3d6000fd5b505050506040513d60208110156107c557600080fd5b8101908080519060200190929190505050905060003073ffffffffffffffffffffffffffffffffffffffff163190506000610811600354600760019054906101000a900460ff166116c6565b9050808203915060008114610999576000600e81111561082d57fe5b600760019054906101000a900460ff16600e81111561084857fe5b1461092b5760008473ffffffffffffffffffffffffffffffffffffffff166329e2caad836040518263ffffffff1660e01b81526004018082815260200191505060206040518083038186803b1580156108a057600080fd5b505afa1580156108b4573d6000803e3d6000fd5b505050506040513d60208110156108ca57600080fd5b8101908080519060200190929190505050905080820391508373ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015610928573d6000803e3d6000fd5b50505b600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015610993573d6000803e3d6000fd5b50610aac565b6000600e8111156109a657fe5b600760019054906101000a900460ff16600e8111156109c157fe5b14610aab5760008473ffffffffffffffffffffffffffffffffffffffff166329e2caad6109ec61124b565b6040518263ffffffff1660e01b81526004018082815260200191505060206040518083038186803b158015610a2057600080fd5b505afa158015610a34573d6000803e3d6000fd5b505050506040513d6020811015610a4a57600080fd5b8101908080519060200190929190505050905080830392508373ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015610aa8573d6000803e3d6000fd5b50505b5b6000829050818303925060008114610b27576000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f19350505050158015610b25573d6000803e3d6000fd5b505b6003600760006101000a81548160ff02191690836004811115610b4657fe5b02179055505050505050565b600062015180905090565b60045481565b6000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614610c08576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260238152602001806126bb6023913960400191505060405180910390fd5b60006004811115610c1557fe5b600760009054906101000a900460ff166004811115610c3057fe5b14610ca3576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252601f8152602001807f636f6e747261637420737461747573206e65656420746f206265206f70656e0081525060200191505060405180910390fd5b6004600760006101000a81548160ff02191690836004811115610cc257fe5b02179055506000809054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc3073ffffffffffffffffffffffffffffffffffffffff16319081150290604051600060405180830381858888f19350505050158015610d45573d6000803e3d6000fd5b50565b600062015180905090565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff1614610df9576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260288152602001806127866028913960400191505060405180910390fd5b42610e02610b52565b6005540110610e5c576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260448152602001806126076044913960600191505060405180910390fd5b42610e65610fc1565b600554011015610ec0576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252604381526020018061264b6043913960600191505060405180910390fd5b60016004811115610ecd57fe5b600760009054906101000a900460ff166004811115610ee857fe5b14610f3e576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252602d81526020018061268e602d913960400191505060405180910390fd5b80600760016101000a81548160ff0219169083600e811115610f5c57fe5b02179055506002600760006101000a81548160ff02191690836004811115610f8057fe5b021790555050565b600760019054906101000a900460ff1681565b600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b60006203f480905090565b600062054600905090565b60055481565b60065481565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b600260009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff163373ffffffffffffffffffffffffffffffffffffffff16146110af576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252602881526020018061259b6028913960400191505060405180910390fd5b426110b8610fc1565b6005540110611112576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260428152602001806127446042913960600191505060405180910390fd5b4261111b610fcc565b600554011015611176576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260448152602001806125c36044913960600191505060405180910390fd5b6002600481111561118357fe5b600760009054906101000a900460ff16600481111561119e57fe5b146111f4576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252602e815260200180612716602e913960400191505060405180910390fd5b80600760016101000a81548160ff0219169083600e81111561121257fe5b021790555050565b6000680ad78ebc5ac620000080838161122f57fe5b04029050919050565b600760009054906101000a900460ff1681565b6000683635c9adc5dea00000905090565b61126461124b565b3410156112d9576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252601e8152602001807f696e73756666696369656e742066756e6420666f72207072656d69756d21000081525060200191505060405180910390fd5b426112e2610d48565b600554031161133c576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004018080602001828103825260388152602001806126de6038913960400191505060405180910390fd5b6000600481111561134957fe5b600760009054906101000a900460ff16600481111561136457fe5b146113d7576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252601f8152602001807f636f6e747261637420737461747573206e65656420746f206265206f70656e0081525060200191505060405180910390fd5b600115156113e4826118ec565b151514611459576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040180806020018281038252601a8152602001807f646f206e6f7420737570706f7274207468697320666c6967687400000000000081525060200191505060405180910390fd5b33600160006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff16021790555060003490506114a761124b565b81039050600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167f17c2f3e0d2d7ead2d68931154e83b5320959605ec7397093a1bd845a60e29c4761150d61124b565b6040518082815260200191505060405180910390a2600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff166108fc829081150290604051600060405180830381858888f1935050505015801561158a573d6000803e3d6000fd5b5081600481905550600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167fd6457c06d7a832c2724560b589156fc0a34d1eea0aae901b1e12d5e279a8f99a826040518082815260200191505060405180910390a2600160009054906101000a900473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff167f596bcc7183bdee7c7371c9072d6e4c6b30dea4f278756da86ed4ec7e471c484030604051808273ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200191505060405180910390a26001600760006101000a81548160ff021916908360048111156116bd57fe5b02179055505050565b60006008600e8111156116d557fe5b82600e8111156116e157fe5b148061170357506009600e8111156116f557fe5b82600e81111561170157fe5b145b806117245750600a600e81111561171657fe5b82600e81111561172257fe5b145b806117455750600b600e81111561173757fe5b82600e81111561174357fe5b145b806117665750600c600e81111561175857fe5b82600e81111561176457fe5b145b806117875750600d600e81111561177957fe5b82600e81111561178557fe5b145b806117a75750600e8081111561179957fe5b82600e8111156117a557fe5b145b156117be57600183816117b657fe5b0490506118e6565b6004600e8111156117cb57fe5b82600e8111156117d757fe5b14156117ef57600883816117e757fe5b0490506118e6565b6005600e8111156117fc57fe5b82600e81111561180857fe5b1415611820576006838161181857fe5b0490506118e6565b6006600e81111561182d57fe5b82600e81111561183957fe5b1415611851576003838161184957fe5b0490506118e6565b6007600e81111561185e57fe5b82600e81111561186a57fe5b1415611882576002838161187a57fe5b0490506118e6565b6002600e81111561188f57fe5b82600e81111561189b57fe5b14156118b357600183816118ab57fe5b0490506118e6565b6000600e8111156118c057fe5b82600e8111156118cc57fe5b14156118e1576118da61124b565b90506118e6565b600090505b92915050565b60007f415320323636000000000000000000000000000000000000000000000000000082148061193b57507f415320333236000000000000000000000000000000000000000000000000000082145b8061196557507f415320333430000000000000000000000000000000000000000000000000000082145b8061198f57507f415320333633000000000000000000000000000000000000000000000000000082145b806119b957507f415320333735000000000000000000000000000000000000000000000000000082145b806119e357507f415320373833000000000000000000000000000000000000000000000000000082145b80611a0d57507f415320363237000000000000000000000000000000000000000000000000000082145b80611a3757507f415320363239000000000000000000000000000000000000000000000000000082145b80611a6157507f415320363537000000000000000000000000000000000000000000000000000082145b80611a8b57507f415320363635000000000000000000000000000000000000000000000000000082145b80611ab557507f415320373836000000000000000000000000000000000000000000000000000082145b80611adf57507f415320313032320000000000000000000000000000000000000000000000000082145b80611b0957507f415320313032340000000000000000000000000000000000000000000000000082145b15611b175760019050612595565b7f5541203231330000000000000000000000000000000000000000000000000000821480611b6457507f554120323935000000000000000000000000000000000000000000000000000082145b80611b8e57507f554120343937000000000000000000000000000000000000000000000000000082145b80611bb857507f554120353335000000000000000000000000000000000000000000000000000082145b80611be257507f554120353737000000000000000000000000000000000000000000000000000082145b80611c0c57507f554120353833000000000000000000000000000000000000000000000000000082145b80611c3657507f554120373533000000000000000000000000000000000000000000000000000082145b80611c6057507f554120383430000000000000000000000000000000000000000000000000000082145b80611c8a57507f554120313235370000000000000000000000000000000000000000000000000082145b80611cb457507f554120313438330000000000000000000000000000000000000000000000000082145b80611cde57507f554120313532360000000000000000000000000000000000000000000000000082145b80611d0857507f554120313538340000000000000000000000000000000000000000000000000082145b80611d3257507f554120313739360000000000000000000000000000000000000000000000000082145b80611d5c57507f554120313834380000000000000000000000000000000000000000000000000082145b80611d8657507f554120313937380000000000000000000000000000000000000000000000000082145b80611db057507f554120323030360000000000000000000000000000000000000000000000000082145b80611dda57507f554120323034340000000000000000000000000000000000000000000000000082145b80611e0457507f554120323036350000000000000000000000000000000000000000000000000082145b80611e2e57507f554120323038300000000000000000000000000000000000000000000000000082145b80611e5857507f554120323136300000000000000000000000000000000000000000000000000082145b80611e8257507f554120323233390000000000000000000000000000000000000000000000000082145b80611eac57507f554120323331390000000000000000000000000000000000000000000000000082145b15611eba5760019050612595565b7f4141203136000000000000000000000000000000000000000000000000000000821480611f0757507f414120373600000000000000000000000000000000000000000000000000000082145b80611f3157507f414120313634000000000000000000000000000000000000000000000000000082145b80611f5b57507f414120313636000000000000000000000000000000000000000000000000000082145b80611f8557507f414120313737000000000000000000000000000000000000000000000000000082145b80611faf57507f414120313739000000000000000000000000000000000000000000000000000082145b80611fd957507f414120323334000000000000000000000000000000000000000000000000000082145b8061200357507f414120323736000000000000000000000000000000000000000000000000000082145b8061202d57507f414120323330350000000000000000000000000000000000000000000000000082145b8061205757507f414120323635320000000000000000000000000000000000000000000000000082145b156120655760019050612595565b7f42362031350000000000000000000000000000000000000000000000000000008214806120b257507f423620313600000000000000000000000000000000000000000000000000000082145b806120dc57507f423620313637000000000000000000000000000000000000000000000000000082145b8061210657507f423620313638000000000000000000000000000000000000000000000000000082145b8061213057507f423620343135000000000000000000000000000000000000000000000000000082145b8061215a57507f423620343136000000000000000000000000000000000000000000000000000082145b8061218457507f423620353136000000000000000000000000000000000000000000000000000082145b806121ae57507f423620363135000000000000000000000000000000000000000000000000000082145b806121d857507f423620363136000000000000000000000000000000000000000000000000000082145b8061220257507f423620363639000000000000000000000000000000000000000000000000000082145b8061222c57507f423620363730000000000000000000000000000000000000000000000000000082145b8061225657507f423620393135000000000000000000000000000000000000000000000000000082145b8061228057507f423620393136000000000000000000000000000000000000000000000000000082145b806122aa57507f423620313431350000000000000000000000000000000000000000000000000082145b806122d457507f423620313531360000000000000000000000000000000000000000000000000082145b806122fe57507f423620313731350000000000000000000000000000000000000000000000000082145b1561230c5760019050612595565b7f444c20343236000000000000000000000000000000000000000000000000000082148061235957507f444c20343330000000000000000000000000000000000000000000000000000082145b8061238357507f444c20343930000000000000000000000000000000000000000000000000000082145b806123ad57507f444c20363130000000000000000000000000000000000000000000000000000082145b806123d757507f444c20363433000000000000000000000000000000000000000000000000000082145b8061240157507f444c20383638000000000000000000000000000000000000000000000000000082145b8061242b57507f444c20393336000000000000000000000000000000000000000000000000000082145b8061245557507f444c20313534380000000000000000000000000000000000000000000000000082145b8061247f57507f444c20313835390000000000000000000000000000000000000000000000000082145b806124a957507f444c20323637300000000000000000000000000000000000000000000000000082145b156124b75760019050612595565b7f484120323335330000000000000000000000000000000000000000000000000082148061250457507f484120323335320000000000000000000000000000000000000000000000000082145b8061252e57507f484120323336300000000000000000000000000000000000000000000000000082145b8061255857507f484120323336310000000000000000000000000000000000000000000000000082145b8061258257507f484120323336320000000000000000000000000000000000000000000000000082145b156125905760019050612595565b600090505b91905056fe6f6e6c7920726567697374726564206f7261636c652063616e206368616e676520726573756c742173686f756c64206368616e6765206265666f726520287363686564756c655f74616b655f6f66665f74696d65202b20656e645f646973707574655f696e74657276616c2973686f756c64207265706f727420616674657220287363686564756c655f74616b655f6f66665f74696d65202b2073746172745f7265706f72745f696e74657276616c2973686f756c64207265706f7274206265666f726520287363686564756c655f74616b655f6f66665f74696d65202b20656e645f7265706f72745f696e74657276616c29636f6e747261637420737461747573206e65656420746f2062652077616974696e6720666f72207265706f72746f6e6c79206f776e65722063616e2063616e63656c207468697320636f6e747261637473686f756c64206275792067657446726f7a656e506572696f64206265666f7265207363686564756c652074616b65206f66662074696d65636f6e747261637420737461747573206e65656420746f2062652077616974696e6720666f72206469737075746573686f756c64206368616e676520616674657220287363686564756c655f74616b655f6f66665f74696d65202b20656e645f7265706f72745f696e74657276616c296f6e6c7920726567697374726564206f7261636c652063616e207265706f727420726573756c742173686f756c6420736574746c6520616674657220287363686564756c655f74616b655f6f66665f74696d65202b20656e645f646973707574655f696e74657276616c29a265627a7a72315820612ed13838b4a948337cce571ea68c9fc0e970d7cf3fb2271f2606a6dc6b4f4264736f6c634300050b00327265616c206d61782062656e656669742073686f756c6420657175616c20746f206d61782062656e6566697421696e73756666696369656e7420696e697469616c206465706f7369742066756e64217363686564756c652074616b65206f66662074696d652073686f756c64206174206c65617374206d61782062656e656669742073686f756c64206c6172676572207468616e207072656d69756d00000000000000000000000000000000000000000000021e19e0c9bab2400000323031392d31302d313700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005da9004400000000000000000000000079f3574d440c21029eecb8e9ee66b23b26cbcb38")
+	_constantinopleOpCodeContract, _ = hex.DecodeString("608060405234801561001057600080fd5b506104d5806100206000396000f3fe608060405234801561001057600080fd5b50600436106100885760003560e01c806381ea44081161005b57806381ea440814610101578063a91b336214610159578063c2bc2efc14610177578063f5eacece146101cf57610088565b80635bec9e671461008d57806360fe47b1146100975780636bc8ecaa146100c5578063744f5f83146100e3575b600080fd5b6100956101ed565b005b6100c3600480360360208110156100ad57600080fd5b8101908080359060200190929190505050610239565b005b6100cd610270565b6040518082815260200191505060405180910390f35b6100eb6102b3565b6040518082815260200191505060405180910390f35b6101436004803603602081101561011757600080fd5b81019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506102f6565b6040518082815260200191505060405180910390f35b61016161036a565b6040518082815260200191505060405180910390f35b6101b96004803603602081101561018d57600080fd5b81019080803573ffffffffffffffffffffffffffffffffffffffff1690602001909291905050506103ad565b6040518082815260200191505060405180910390f35b6101d761045f565b6040518082815260200191505060405180910390f35b5b60011561020b5760008081548092919060010191905055506101ee565b7f8bfaa460932ccf8751604dd60efa3eafa220ec358fccb32ef703f91c509bc3ea60405160405180910390a1565b80600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a250565b6000805460081d905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b6000805460081c905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b60008073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141561033157600080fd5b813f9050807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a2809050919050565b6000805460081b905080600081905550807fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a280905090565b60008073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff1614156103e857600080fd5b7fbde7a70c2261170a87678200113c8e12f82f63d0a1d1cfa45681cbac328e87e382600054604051808373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020018281526020019250505060405180910390a16000549050919050565b60008080602060406000f59150817fdf7a95aebff315db1b7716215d602ab537373cdb769232aae6055c06e798425b60405160405180910390a2819150509056fea265627a7a72305820209a8ef04c4d621759f34878b27b238650e8605c8a71d6efc619a769a64aa9cc64736f6c634300050a0032")
 )
